Apply button classes to drafts by hand
If we let jquery mobile enhance this button then you get a bug whereby
once you've gone to the drafts screen and returned moving the map cause
the hover code of the button to fire. I can't work out why that happens
and this solves the problem at the cost of the highlighting of the
button not working.

Fixes #4