Fix race condition when setting currentView
As the front page only serves to work out what to do on load we should
never set the currentView to front-page. However, due to execution order
what actually happens is that the navigation to the front page completes
after the navigation to the actual page displayed. This means that the
incorrect currentView is set and the easiest way to fix is to prevent
the currentView being set to the front page