Allow details screen to scroll on Android
This fixes #113 in that it means the user can move the screen to see
what fields are available. Previously this was prevented by setting the
adjustPan mode for the soft input mode which had fixed a bug with the
keyboard not dismissing on transition. This has been resolved by adding a
plugin for Android to control display of the soft keyboard and hiding
the keyboard before navigation in the navigate method of FMSView.

Sadly this then meant that the screen size was being measured *before*
the keyboard was dismissed with the result that the screen transitioned
too was incorrectly displayed. This was resolved by taking the screen
height on start up, storing that as a property of the FMS global object
and then using this property instead of measuring the screen size at the
time of transition.

The map has also been expanded to fill the whole screen rather than
stopping at the bottom of the navigation bar as on Android when you
scroll the screen the map does not scroll and the gap at the top was
visible.