Upgrade to Cordova 6
This updates the whole app (ios and android) to use Cordova 6
and the latest android/ios platforms and plugins. In addition
I've added the necessary new icons and splash screens, as well
as updating the whitelist to work with AppTransportSecurity in
iOS9. I've also updated the instructions in the README to be
more useful.