Add a hook to remove unnecessary permissions
The cordova media plugin asks for lots of permissions but
we don't actually need them all (it could be used to film
video, record audio or play files, but we don't do those
things), so they can be removed. It's slightly hacky to do
so because Cordova doesn't offer a built-in way, so instead
we add an after_prepare hook to remove them from the right
file.