Tidy up how custom icons are included in css
This tidies up the custom theme's icons, which were overly complicated and
badly setup - almost as if we were (ab)using css intended to make extra large
icons to make retina icons instead, but we'd only got halfway. I've tidied
this up so that we use the 36px icons properly - keeping most properties the
same as the 18px ones, but scaling the background image appropriately to make
it hi-dpi. Since there seemed to be some issue with iOS not obeying the media
query on the @import statement that loaded this css, I also decided to just
bundle the rules directly into fms-jquerymobile.css

In addition, I took the opportunity to make black icons the default (instead
of white), meaning that we could remove an awful lot of css overrides in
fms.css that were swapping colors for different icons individually.

Closes #217