Upgrade to latest Cordova platforms
This requires dropping support for iOS 8 and lower, but given that
version’s single-digit market share this won’t be a problem.
Existing installs of the app on iOS 8 will continue to function.

Recent versions of Cordova require package.json and package-lock.json
files in addition to config.xml, so example files have been added.