Process index.html as template at build time
There are a few places in index.html that depend on CONFIG in order to load
the right cobrand stylesheet/language/map etc.
This commit removes the run time script loading and instead processes
index.html as a lodash template at build time to pull in the correct stylesheets
and scripts.