Show error to user if registration password is rejected by server
This occurs if the password is too short or too common, for example.

The user's previously entered password is remembered in the UI. The flow
here is not ideal - it would be better to show the error on the same
screen as the password field, but that would require a bit more
reworking of the code.