Disable report form on a per-field-value basis
This commit adds a 'start new report' button which is shown when the
user returns to the app after navigating away to an external link.

We don't want to just discard their report as soon as they navigate away,
like we do with the disabled category behaviour, because by the time
they reach this screen they have likely entered a fair bit of data.