Show unresponsive message immediately after location is confirmed
If the user is making a report in a location where the body refuses
all reports, show them a message to this effect immediately after they
have confirmed the report location and before they've started entering
report details & photo.

Technically this adds a new view before the photo view which is shown
with the contents of `top_message` from `/report/new/ajax`, if present -
so it doesn't have to only be used for unresponsive messages if we
change what the server returns in the future.