Fix bug preventing moving past the photo screen in some situations
When returning to the map after partially completing a report, the next
report you made would crash on the "next" button of the photo screen.

This was caused by AroundView keeping a reference to the incorrect model
object, meaning FMS.currentDraft was an empty object with no categories
to be shown by DetailView.

Fixes #244.